Verify the execution compatibility in start
