+++ packages/prover/src/utils/execution.ts
@@ -1,100 +1,10 @@
 import {Common, CustomChain, Hardfork} from "@ethereumjs/common";
-import {Logger} from "@lodestar/logger";
-import {ELRequestHandler} from "../interfaces.js";
-import {
-  ELApi,
-  ELApiParams,
-  ELApiReturn,
-  ELTransaction,
-  JsonRpcRequest,
-  JsonRpcResponse,
-  JsonRpcResponseWithResultPayload,
-  JsonRpcBatchRequest,
-  JsonRpcBatchResponse,
-} from "../types.js";
-import {
-  isRequest,
-  isValidBatchResponse,
-  isValidResponse,
-  logRequest,
-  logResponse,
-  mergeBatchReqResp,
-} from "./json_rpc.js";
-import {isBlockNumber, isNullish, isPresent} from "./validation.js";
+import {ELApiParams, ELApiReturn, ELTransaction} from "../types.js";
+import {isValidResponse} from "./json_rpc.js";
+import {isBlockNumber, isPresent} from "./validation.js";
+import {ELRpc} from "./rpc.js";
 
 export type Optional<T, K extends keyof T> = Omit<T, K> & {[P in keyof T]?: T[P] | undefined};
-
-export class ELRpc {
-  private handler: ELRequestHandler;
-  private logger: Logger;
-
-  constructor(handler: ELRequestHandler, logger: Logger) {
-    this.handler = handler;
-    this.logger = logger;
-  }
-
-  async request<K extends keyof ELApi, E extends boolean>(
-    method: K,
-    params: ELApiParams[K],
-    opts: {raiseError: E}
-  ): Promise<E extends false ? JsonRpcResponse<ELApiReturn[K]> : JsonRpcResponseWithResultPayload<ELApiReturn[K]>> {
-    const {raiseError} = opts;
-
-    const payload: JsonRpcRequest = {jsonrpc: "2.0", method, params, id: this.getRequestId()};
-    logRequest(payload, this.logger);
-
-    const response = await this.handler(payload);
-    logResponse(response, this.logger);
-
-    if (raiseError && !isValidResponse(response)) {
-      throw new Error(`Invalid response from RPC. method=${method} params=${JSON.stringify(params)}`);
-    }
-
-    return response as JsonRpcResponseWithResultPayload<ELApiReturn[K]>;
-  }
-
-  async batchRequest<E extends boolean>(
-    input: JsonRpcBatchRequest,
-    opts: {raiseError: E}
-  ): Promise<
-    E extends false
-      ? {request: JsonRpcRequest; response: JsonRpcResponse}[]
-      : {request: JsonRpcRequest; response: JsonRpcResponseWithResultPayload<unknown>}[]
-  > {
-    const payloads: JsonRpcBatchRequest = [];
-
-    for (const req of input) {
-      if (isRequest(req) && isNullish(req.id)) {
-        payloads.push({jsonrpc: "2.0", method: req.method, params: req.params, id: this.getRequestId()});
-      } else {
-        payloads.push(req);
-      }
-    }
-
-    logRequest(payloads, this.logger);
-    const response = await this.handler(payloads);
-    logResponse(response, this.logger);
-
-    if (isNullish(response)) {
-      throw new Error("Invalid empty response from server.");
-    }
-
-    if (opts.raiseError && !isValidBatchResponse(payloads, response as JsonRpcBatchResponse)) {
-      throw new Error(
-        `Invalid response from RPC. payload=${JSON.stringify(payloads)} response=${JSON.stringify(response)}}`
-      );
-    }
-
-    return mergeBatchReqResp(payloads, response as JsonRpcBatchResponse) as E extends false
-      ? {request: JsonRpcRequest; response: JsonRpcResponse}[]
-      : {request: JsonRpcRequest; response: JsonRpcResponseWithResultPayload<unknown>}[];
-  }
-
-  private getRequestId(): string {
-    // TODO: Find better way to generate random id
-    return (Math.random() * 10000).toFixed(0);
-  }
-}
 
 export async function getELCode(rpc: ELRpc, args: ELApiParams["eth_getCode"]): Promise<ELApiReturn["eth_getCode"]> {
   const codeResult = await rpc.request("eth_getCode", args, {raiseError: false});
